Adapt the CI workflow to our needs
Firstly, clear out scripts and makefiles we don't need right now.

The terraform linting doesn't work, and the preprovided makefile demands
a lot of tasks are defined, so I've stripped out a lot of it and just
implemented the essentials to get the pipeline working.

Keep .tool-versions and pre-commit, but just run a secrets scan for now.
We will add in stuff as we go - e.g. python linters.

In CI:
- code scanning and tests can be parallelised
- we're better off without a push trigger on every branch, so we don't get duplicate jobs triggered
  for every PR (one for pull_request, one for push). We will be
  configuring the repo to prevent direct pushes to the main branch.