Automatically publish images to the registry
We are using Github packages to publish our images publicly.
We don't yet have continuous deployment, but we will pull
from that registry as part of the terraform apply.

This sets up Github actions to publish the image from the github
actions workflow, as documented here:
https://docs.github.com/en/packages/managing-github-packages-using-github-actions-workflows/publishing-and-installing-a-package-with-github-actions#publishing-a-package-using-an-action

Authentication is via the special GITHUB_ACTIONS secret, as documented
here:
https://docs.github.com/en/actions/security-for-github-actions/security-guides/automatic-token-authentication

The registry matches the repository name:
https://github.com/NHSDigital/manage-breast-screening/pkgs/container/manage-breast-screening