Limit docker tags to git shas, branches and tags
DOCKER_METADATA_PR_HEAD_SHA ensures we use the sha of the branch head
when the build is triggered by pull request, rather than an ephermeral
merge commit.

This is not strictly what was built, but ensures you can check out the
code locally if there is a problem with the image.