Add more custom filters
Add the following custom Jinja filters form the prototype
- formatNhsNumber
- sentenceCase
- formateRelativeDate
- formatTimeString

We've decided to keep the camelcase in the templates, but use normal
python snake case in the code.

This should make it a bit easier to adapt existing prototype templates
as we build out the app.

Since these filters are easy documented by example, I've added examples
in all the docstrings, and enabled doctest to include those in our test
suite.

For the relative dates, I'm using a simplified version of the logic in
the prototype, that only handles the near future/past.

I tried using django.contrib.humanize, and arrow's humanize method, but
neither of these produced output consistent with our style guide.
Humanizing dates in the more distant past/future (a year ago, a month
ago etc) are not essential for what we're working on right now, so I'm
leaving it for another day.