Add appointment checkin button
We've decided to style the button as a link here (following the
prototype), because we don't currently have a good button component
that we can use instead - the secondary button variant is too large
and obtrusive.

I considered using the small variant from the
manage vaccinations in schools prototype[^1], but decided against it,
as it looked weird next to the status tag.

We can't use a real link here because clicking on it performs a
non-reversible action - we need this to be a POST action.

Visually styling buttons as links is an imperfect solution; for example
it can cause confusion if somebody describes the component to a screen reader
user who doesn't see it as a link. However, this is a first iteration
and the design will probably change.

^[1]: https://github.com/nhsuk/manage-vaccinations-in-schools-prototype/blob/8f088730d35a2d2d44ea0f1dc7ae2b5096bea4d7/app/assets/stylesheets/_button.scss#L106