Replace sass with sass-embedded
> This package is an alternative to the sass package. It supports the
same JS API as sass and is maintained by the same team, but where the
sass package is pure JavaScript, sass-embedded is instead a JavaScript
wrapper around a native Dart executable. This means sass-embedded will
generally be much faster especially for large Sass compilations, but it
can only be installed on the platforms that Dart supports: Windows, Mac
OS, and Linux.

https://www.npmjs.com/package/sass-embedded