Add ethnic group choices to participant models
This replicates the classification we have in the prototype, and aligns
the participant summary list with the one in the prototype.

The list comes from the GOV.UK design system pattern[^1] which uses the
standard categorisation from the ONS, based on the 2021 census. The
GOV.UK formatting is slightly changed from the ONS guidance.

A "Prefer not to say" option is included so that we can remember
if somebody entered this (as opposed to the data just not being
available).

[^1]: https://design-system.service.gov.uk/patterns/equality-information/