Remove the PR template from the repo
This is not currently adding any value for our team, and it interferes
with the default github behaviour of using the commit message for single
commit PRs.

This feels like it would be more useful for external contributions, but
we are not expecting to have any of those for this repo.

The checklist just reinforces basic expectations we should already be
following, and will be checked as part of the required code review.
Code style will be enforced by linters soon, and we will also be setting
up Sonar Cloud.

Also docs/contributing.md is a dead link and nobody has noticed.