Adopt the template packaged with NHS.UK frontend
This takes care of a lot of things for us, so we just need to change
what's custom. I've removed all the footer items and header items
that don't do anything yet. I think it's better to add these
when they're functional so the app doesn't appear more functional than
it is.