Use snake case consistently in templates
We've opted to follow the convention of using snake_case in our jinja
templates for any macros and blocks we write ourselves. Python logic appearing
in templates should also be in snake_case. This should, in practice,
mean that the only camelCase in our templates comes from external
libraries.

Update our templates to follow these conventions.