Remove broken workaround for summary list
We've been passing an empty object as an action value, to workaround
an issue the row borders don't get rendered properly when some rows
have actions and some don't.

However, this workaround causes accessibility issues as it generates
broken html with invisible links in, i.e.

    <dd class="nhsuk-summary-list__actions">
        <a href="">
            </a>
    </dd>

The issue with the borders was fixed in https://github.com/nhsuk/nhsuk-frontend/pull/1324
and so we will pick it up when it's release and we've pulled the fix into nhsuk-frontend-jinja.

In the meantime, strip out the unwanted html, accepting that the borders
will look weird until the we've updated to the fixed version.