Update to the latest version of Axe
axe_playwright_python allows you to pass in the script to use for Axe,
but when I tried this I got a cryptic error from the `page.evaluate`
call that loads the script:

```python
        # inject `Axe` into document
        page.evaluate(self.axe_script)
```

    E           playwright._impl._errors.Error: Page.evaluate: SyntaxError: Invalid regular expression flags
    E               at eval (<anonymous>)
    E               at UtilityScript.evaluate (<anonymous>:313:29)
    E               at UtilityScript.<anonymous> (<anonymous>:1:44)

I don't know how `page.evaluate` works, but the docs[^1] talk about
javascript *expressions* rather than entire scripts, so I'm wondering if
the library code shouldn't be using `page.add_script_tag` or
`page.add_init_script` instead of `page.evalute`. There is an issue in
playright-python[^2] that uses the former, and both worked without
issues when I tested it.

As a workaround, I've implemented a simple class that enables us to
inject Axe into the page ourselves, which just uses the
axe_playright_python library for its AxeResult object, which does the
helpful formatting.

[^1]: https://playwrighat.dev/python/docs/api/class-page#page-evaluate
[^2]: https://github.com/microsoft/playwright-python/issues/2266#issuecomment-1914376817